refactor(auth): move reusable variables and components to entities, and move icons to shared
