+++ src/entities/auth/index.ts
@@ -3,3 +3,5 @@ export type * as TAuth from './model/types';
 export * as CAuth from './model/const';
 export { AuthHttp } from './api/http';
 export { AuthQueries } from './api/queries';
+export { type OAuthProviderMeta } from './model/const';
+export { OAuthButton } from './ui/OAuthButton';
