--- src/features/auth/oauth-login/ui/OAuthIcons.tsx
@@ -1,56 +1,0 @@
-type IconProps = { className?: string };
-
-export function GoogleIcon({ className }: IconProps) {
-  return (
-    <svg className={className} xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24">
-      <path
-        d="M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92c-.26 1.37-1.04 2.53-2.21 3.31v2.77h3.57c2.08-1.92 3.28-4.74 3.28-8.09z"
-        fill="#4285F4"
-      />
-      <path
-        d="M12 23c2.97 0 5.46-.98 7.28-2.66l-3.57-2.77c-.98.66-2.23 1.06-3.71 1.06-2.86 0-5.29-1.93-6.16-4.53H2.18v2.84C3.99 20.53 7.7 23 12 23z"
-        fill="#34A853"
-      />
-      <path
-        d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l2.85-2.22.81-.62z"
-        fill="#FBBC05"
-      />
-      <path
-        d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z"
-        fill="#EA4335"
-      />
-      <path d="M1 1h22v22H1z" fill="none" />
-    </svg>
-  );
-}
-
-export function YandexIcon({ className }: IconProps) {
-  return (
-    <svg className={className} viewBox="0 0 48 48" xmlns="http://www.w3.org/2000/svg">
-      <path
-        fill="#fc3f1d"
-        d="M26.9 13.3v24.2h4.9v-28h-7.1c-7.1 0-10.9 3.7-10.9 9.1 0 4.3 2.1 6.8 5.7 9.5l-6.4 9.5h5.3L25.5 27 23 25.3c-3-2-4.4-3.6-4.4-7 0-3 2.1-5 6.1-5h2.2z"
-      />
-    </svg>
-  );
-}
-export function GithubIcon({ className }: IconProps) {
-  return (
-    <svg className={className} viewBox="0 0 48 48" xmlns="http://www.w3.org/2000/svg">
-      <path
-        fill="#fff"
-        d="M17.6 46.7c0 .7-.5 1.3-1.6 1.3h16.1c-1.5 0-1.9-.6-1.9-1.3v-6.9c0-2.4-.8-3.9-1.7-4.7 5.6-.6 11.5-2.8 11.5-12.5 0-2.8-1-5-2.6-6.8.2-.6 1.1-3.2-.2-6.7 0 0-2.1-.7-6.9 2.6-2-.6-4.2-.8-6.3-.8s-4.2.3-6.3.8c-4.8-3.3-6.9-2.6-6.9-2.6-1.4 3.5-.5 6.1-.2 6.7C9 17.6 8 19.8 8 22.6c0 9.7 5.9 11.9 11.4 12.5-.7.6-1.4 1.8-1.6 3.4-1.5.7-5.1 1.8-7.3-2.1 0 0-1.3-2.4-3.9-2.6 0 0-2.5 0-.2 1.6 0 0 1.7.8 2.8 3.7 0 0 1.5 4.5 8.4 3v4.6z"
-      />
-    </svg>
-  );
-}
-export function VkontakteIcon({ className }: IconProps) {
-  return (
-    <svg className={className} viewBox="0 0 48 48" xmlns="http://www.w3.org/2000/svg">
-      <path
-        fill="#fff"
-        d="M25.5 35C14.5 35 8.3 27.5 8 15h5.5c.2 9.2 4.2 13.1 7.4 13.9V15h5.2v7.9c3.2-.3 6.5-3.9 7.6-7.9h5.2c-.9 4.9-4.5 8.5-7 10 2.6 1.2 6.7 4.3 8.2 10h-5.7c-1.2-3.8-4.3-6.7-8.3-7.1V35h-.6z"
-      />
-    </svg>
-  );
-}
