+++ src/shared/ui/icon/auth/GithubIcon.tsx
@@ -1,0 +1,12 @@
+import { IconProps } from '../types';
+
+export function GithubIcon({ className }: IconProps) {
+  return (
+    <svg className={className} viewBox="0 0 48 48" xmlns="http://www.w3.org/2000/svg">
+      <path
+        fill="#fff"
+        d="M17.6 46.7c0 .7-.5 1.3-1.6 1.3h16.1c-1.5 0-1.9-.6-1.9-1.3v-6.9c0-2.4-.8-3.9-1.7-4.7 5.6-.6 11.5-2.8 11.5-12.5 0-2.8-1-5-2.6-6.8.2-.6 1.1-3.2-.2-6.7 0 0-2.1-.7-6.9 2.6-2-.6-4.2-.8-6.3-.8s-4.2.3-6.3.8c-4.8-3.3-6.9-2.6-6.9-2.6-1.4 3.5-.5 6.1-.2 6.7C9 17.6 8 19.8 8 22.6c0 9.7 5.9 11.9 11.4 12.5-.7.6-1.4 1.8-1.6 3.4-1.5.7-5.1 1.8-7.3-2.1 0 0-1.3-2.4-3.9-2.6 0 0-2.5 0-.2 1.6 0 0 1.7.8 2.8 3.7 0 0 1.5 4.5 8.4 3v4.6z"
+      />
+    </svg>
+  );
+}
