+++ src/shared/ui/icon/auth/VkontakteIcon.tsx
@@ -1,0 +1,12 @@
+import { IconProps } from '../types';
+
+export function VkontakteIcon({ className }: IconProps) {
+  return (
+    <svg className={className} viewBox="0 0 48 48" xmlns="http://www.w3.org/2000/svg">
+      <path
+        fill="#fff"
+        d="M25.5 35C14.5 35 8.3 27.5 8 15h5.5c.2 9.2 4.2 13.1 7.4 13.9V15h5.2v7.9c3.2-.3 6.5-3.9 7.6-7.9h5.2c-.9 4.9-4.5 8.5-7 10 2.6 1.2 6.7 4.3 8.2 10h-5.7c-1.2-3.8-4.3-6.7-8.3-7.1V35h-.6z"
+      />
+    </svg>
+  );
+}
