+++ src/shared/ui/icon/index.ts
@@ -1,0 +1,5 @@
+export { type IconProps } from './types';
+export { GithubIcon } from './auth/GithubIcon';
+export { GoogleIcon } from './auth/GoogleIcon';
+export { VkontakteIcon } from './auth/VkontakteIcon';
+export { YandexIcon } from './auth/YandexIcon';
