+++ src/shared/ui/icon/types.ts
@@ -1,0 +1,1 @@
+export type IconProps = { className?: string };
